+++ src/chat/chat-socket.js
@@ -4,14 +4,11 @@ const chat = require('../socket');
 // 로비에 연결 되었을때
 chat.on('connection', async (socket) => {
     console.log(socket.id + ' 여긴 채팅방!!!');
-    // 로비 채팅
-    socket.on('sendLobbyMsg', async (msg) => {
-        console.log(msg);
-        const sids = chat.sockets.adapter.sids;
-        const rooms = chat.sockets.adapter.rooms;
 
-        console.log('sids:::', sids);
-        console.log('rooms:::', rooms);
+    // 로비 채팅
+    socket.on('sendLobbyMsg', async (payload, callback) => {
+        socket.broadcast.emit('receiveLobbyMsg', payload);
+        callback();
     });
     // 룸 채팅
     socket.on('sendRoomMsg', async (payload, roomNum, callback) => {
@@ -20,3 +17,27 @@ chat.on('connection', async (socket) => {
         callback();
     });
 });
+
+function publicRooms() {
+    const sids = chat.sockets.adapter.sids;
+    const rooms = chat.sockets.adapter.rooms;
+    const publicRooms = [];
+    rooms.forEach((_, key) => {
+        if (sids.get(key) === undefined) {
+            publicRooms.push(key);
+        }
+    });
+    return publicRooms;
+}
+
+const getLobbyUsers = async () => {
+    const sids = await chat.sockets.adapter.sids;
+    const rooms = await chat.sockets.adapter.rooms;
+    const lobbyUsers = [];
+    await rooms.forEach((_, key) => {
+        if (sids.get(key) !== undefined && sids.get(key).size === 1) {
+            lobbyUsers.push(key);
+        }
+    });
+    return lobbyUsers;
+};
